Extract LLM library functions
